Refactor screenshot handling into shared utility function
