+++ common/modules/utilities.js
@@ -123,6 +123,58 @@ export async function json(url) {
     return await fetch(url).then(response => response.json()).catch(error => console.error("An error occurred.", error));
 }
 
+export function processScreenshots(app, maxCount = Infinity) {
+    let html = "";
+
+    // New format with support for universal apps
+    if (app.screenshots) {
+        // Helper function to process screenshot array
+        const processScreenshotArray = (screenshots) => {
+            if (!Array.isArray(screenshots) || !screenshots.length) return "";
+
+            return screenshots
+                .slice(0, maxCount)
+                .map((screenshot, i) => {
+                    let imageURL;
+
+                    if (typeof screenshot === "string" && isValidHTTPURL(screenshot)) {
+                        imageURL = screenshot;
+                    } else if (screenshot && typeof screenshot === "object" && screenshot.imageURL) {
+                        imageURL = screenshot.imageURL;
+                    }
+
+                    if (!imageURL) return ""; // Skip invalid entries
+
+                    const altText = `${app.name} screenshot ${i + 1}`;
+                    return `<img src="${imageURL}" alt="${altText}" class="screenshot">`;
+                })
+                .join("");
+        };
+
+        if (Array.isArray(app.screenshots)) {
+            // Standard format: array of strings or objects
+            html = processScreenshotArray(app.screenshots);
+        } else if (typeof app.screenshots === 'object') {
+            // Universal Apps format: object with iphone and ipad keys
+            // Only process iPhone screenshots as requested
+            if (app.screenshots.iphone && Array.isArray(app.screenshots.iphone)) {
+                html = processScreenshotArray(app.screenshots.iphone);
+            }
+        }
+    } else if (app.screenshotURLs) {
+        // Legacy format
+        html = app.screenshotURLs
+            .slice(0, maxCount)
+            .map((url, i) => {
+                if (!url) return "";
+                return `<img src="${url}" alt="${app.name} screenshot ${i + 1}" class="screenshot">`;
+            })
+            .join("");
+    }
+
+    return html;
+}
+
 const $ = selector => selector.startsWith("#") && !selector.includes(".") && !selector.includes(" ")
     ? document.getElementById(selector.substring(1))
     : document.querySelectorAll(selector);
